+++ src/client/components/common/Dropdown/Dropdown.tsx
@@ -4,7 +4,7 @@ import ArrowSVG from './ArrowSVG.svg';
 import classNames from 'classnames';
 import styles from './Dropdown.module.scss';
 import Link from 'next/link';
-import { isLocalLink } from '../../../../shared/utils/link';
+import { isLocalLink } from '../../../../utils/link';
 import Image from 'next/image';
 
 export interface DropdownProps {
