+++ src/utils/locationUtils.test.ts
