+++ src/utils/timeUtils.ts
@@ -53,3 +53,11 @@ export const timeAgoLabel = (dateString: string): string => {
   }
   return `More than 3 months ago`;
 };
+
+export function getTodayDateRange(): [Date, Date] {
+  const timestampStart = new Date();
+  const timestampEnd = new Date();
+  timestampStart.setHours(0, 0, 0, 0);
+  timestampEnd.setHours(23, 59, 59, 59);
+  return [timestampStart, timestampEnd];
+}
